harden stripCreds and share mapLimit across storage routes
