+++ lib/index.js
@@ -10,6 +10,7 @@ module.exports = {
 	password: require("./utils/password.json"),
 	subprocess: require("./utils/subprocess.js"),
 	formatBytes: require("./utils/formatBytes.js"),
+	mapLimit: require("./utils/mapLimit.js"),
 	tempMiddleware: require("./utils/tempMiddleware.js"),
 	jsonFileHanding: require("./utils/jsonFileHandling.js"),
 	randomID: require("./utils/randomID.js"),
