+++ lib/utils/mapLimit.js
@@ -1,0 +1,14 @@
+const mapLimit = async (items, limit, fn) => {
+	const results = new Array(items.length)
+	let next = 0
+	const worker = async () => {
+		while (next < items.length) {
+			const i = next++
+			results[i] = await fn(items[i], i)
+		}
+	}
+	await Promise.all(Array.from({ length: Math.min(limit, items.length) }, worker))
+	return results
+}
+
+module.exports = mapLimit
